python dict: repalce iteritems with items
In py2:

    dict.items() -> list
    dict.iteritems() -> iterator

iteritems() is preferred for less memory usage in py2.

In py3:

    dict.items() -> iterator
    dict.iteritems() -> removed

Use items() instead of iteritems(), to be python 2/3 compatible.

In py2, the returned list may use more memory, but shouldn't be problem
in this case.

Didn't use six, which is unnecessary here and introduce noise in code.

Signed-off-by: Joe Guo <guoqiao@gmail.com>